+++ src/execution/executor_seq_scan.h
@@ -112,8 +112,8 @@ class SeqScanExecutor : public AbstractExecutor {
                 return ai > bi ? 1 : ((ai < bi) ? -1 : 0);
             }
             case TYPE_FLOAT: {
-                float af = *(float *) a;
-                float bf = *(float *) b;
+                double af = *(double *) a;
+                double bf = *(double *) b;
                 return af > bf ? 1 : ((af < bf) ? -1 : 0);
             }
             case TYPE_STRING:
